sdsn_overall_transmission: Include Brazil & China
Update the methodology for sdsn_overall_transmission as recommended by 
Jeffrey Sachs: Any country with more than 100 new cases per million, 
gets a rating of 5 (very high), even when no test data is present. The 
requirement to have testing data only makes sense for the better 
classifications because we don't want to give a country a 2 or 3 just 
because they don't test enough. The classification for China is just 
based on new_cases_per_million, not on the data on tests available.