Update methodology for sdsn_overall_transmission
Use owid_positive_rate rather than sdsn_tests_per_case to calibrate the 
results between (1) suppressed and (2) low. This is for two reasons:
- Using positive test rate ≤ 0.05 rather than tests per case ≥ 20 for 
the suppressed category, because it gets around the issue of incorrect 
missing values when cases are zero and tests per case is equal to 
infinity.
- Using the 7-day average for the positive test rate rather than the 
14-day average. The 7-day average is available from OWID, so we would 
just use those data points. The other option is to calculate a 14-day 
average over the 7-day average, but just using the 7-day average strikes 
us as the better approach.

Using a positive rate of ≤ 5% is also in line with WHO guidelines: 
https://www.who.int/publications/i/item/public-health-criteria-to-adjust-public-health-and-social-measures-in-the-context-of-covid-19